Move jsonpointer to authenticationflow
