+++ pkg/lib/authenticationflow/declarative/intent_signup_flow_step_user_profile.go
@@ -40,7 +40,7 @@ func (*IntentSignupFlowStepUserProfile) Kind() string {
 
 func (i *IntentSignupFlowStepUserProfile) CanReactTo(ctx context.Context, deps *authflow.Dependencies, flows authflow.Flows) (authflow.InputSchema, error) {
 	if len(flows.Nearest.Nodes) == 0 {
-		current, err := flowObject(authflow.GetFlowRootObject(ctx), i.JSONPointer)
+		current, err := authflow.FlowObject(authflow.GetFlowRootObject(ctx), i.JSONPointer)
 		if err != nil {
 			return nil, err
 		}
@@ -60,7 +60,7 @@ func (i *IntentSignupFlowStepUserProfile) CanReactTo(ctx context.Context, deps *
 func (i *IntentSignupFlowStepUserProfile) ReactTo(ctx context.Context, deps *authflow.Dependencies, flows authflow.Flows, input authflow.Input) (*authflow.Node, error) {
 	var inputFillUserProfile inputFillUserProfile
 	if authflow.AsInput(input, &inputFillUserProfile) {
-		current, err := flowObject(authflow.GetFlowRootObject(ctx), i.JSONPointer)
+		current, err := authflow.FlowObject(authflow.GetFlowRootObject(ctx), i.JSONPointer)
 		if err != nil {
 			return nil, err
 		}
